+++ dummyio/requirements.txt
@@ -1,0 +1,1 @@
+granian==1.4.4
