--- helper-images/dummyio/requirements.txt
@@ -1,1 +1,0 @@
-granian
